+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.1.1-pre
-*
+* Modify upgradeDependencies to also uprade dependencies in devDependencies and peerDependencies properties
 
 ### 0.1.0
 * Set min node engine to >= v0.10.0